created base class to declutter example apps
